Win32 nd Win64 compiler warning fixes. This is replacement for r7886 which was undone in r7909. Approved by Mattias Jonsson in RB#5249

+++ storage/innobase/row/row0ins.cc
@@ -3097,7 +3097,8 @@ row_ins_spatial_index_entry_set_mbr_field(
 
 	/* obtain the MBR */
 	rtree_mbr_from_wkb(dptr + GEO_DATA_HEADER_SIZE,
-			   dlen - GEO_DATA_HEADER_SIZE, SPDIMS, mbr);
+			   static_cast<uint>(dlen - GEO_DATA_HEADER_SIZE),
+			   SPDIMS, mbr);
 
 	/* Set mbr as index entry data */
 	dfield_write_mbr(field, mbr);
